Add last seen logic to App
